+++ examples/next-openai/app/api/use-chat-persistence/route.ts
@@ -1,5 +1,5 @@
 import { openai } from '@ai-sdk/openai';
-import { appendResponseMessages, streamText } from 'ai';
+import { appendResponseMessages, createIdGenerator, streamText } from 'ai';
 import { saveChat } from './chat-store';
 
 export async function POST(req: Request) {
@@ -8,6 +8,11 @@ export async function POST(req: Request) {
   const result = streamText({
     model: openai('gpt-4o-mini'),
     messages,
+    // id format for server-side messages:
+    experimental_generateMessageId: createIdGenerator({
+      prefix: 'msgs',
+      size: 16,
+    }),
     async onFinish({ response }) {
       await saveChat({
         id,