+++ packages/cerebras/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/cerebras
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [00114c5]
+  - @ai-sdk/provider-utils@2.0.8
+  - @ai-sdk/openai-compatible@0.0.18
+
 ## 0.0.2
 
 ### Patch Changes