+++ packages/replicate/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/replicate
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [00114c5]
+  - @ai-sdk/provider-utils@2.0.8
+
 ## 0.0.3
 
 ### Patch Changes