+++ packages/swarm/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/swarm
 
+## 0.0.49
+
+### Patch Changes
+
+- Updated dependencies [33592d2]
+  - ai@4.0.40
+
+## 0.0.48
+
+### Patch Changes
+
+- Updated dependencies [00114c5]
+- Updated dependencies [00114c5]
+  - ai@4.0.39
+
 ## 0.0.47
 
 ### Patch Changes