+++ packages/ui-utils/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/ui-utils
 
+## 1.0.12
+
+### Patch Changes
+
+- 33592d2: fix (ai/core): switch to json schema 7 target for zod to json schema conversion
+
+## 1.0.11
+
+### Patch Changes
+
+- 00114c5: feat (ui): generate and forward message ids for response messages
+- Updated dependencies [00114c5]
+  - @ai-sdk/provider-utils@2.0.8
+
 ## 1.0.10
 
 ### Patch Changes